+++ crates/api-desc/src/radio.rs
@@ -14,6 +14,8 @@
 
 use crate::*;
 
+mod ble;
+
 pub(crate) fn new() -> Item {
     let docs = docs! {
         /// Radio operations.
@@ -21,42 +23,12 @@ pub(crate) fn new() -> Item {
     let name = "radio".into();
     let items = vec![
         item! {
-            /// Reads radio packet into a buffer.
-            fn read "rr" {
-                /// Address of the buffer.
-                ptr: *mut u8,
-
-                /// Length of the buffer in bytes.
-                len: usize,
-            } -> {
-                /// Number of bytes read (or negative value for errors).
-                ///
-                /// This function does not block and may return zero.
-                len: isize,
-            }
-        },
-        item! {
-            /// Register a handler for radio events.
-            fn register "re" {
-                /// Function called on radio events.
-                ///
-                /// The function takes its opaque `data` as argument.
-                handler_func: fn { data: *mut u8 },
-
-                /// The opaque data to use when calling the handler function.
-                handler_data: *mut u8,
-            } -> {}
-        },
-        item! {
-            /// Unregister handlers for radio events.
-            fn unregister "rd" {} -> {}
-        },
-        item! {
             /// Describes errors on radio operations.
             enum Error {
                 Unknown = 0,
             }
         },
+        ble::new(),
     ];
     Item::Mod(Mod { docs, name, items })
 }